requirements.txt
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Markdown==3.8
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
djangorestframework-simplejwt==5.5.0
django-widget-tweaks
daphne
django-redis
channels_redis
graphene-django
python-memcached==1.62
pymemcache
django-oauth-toolkit
celery
psycopg2-binary
django-allauth
google-api-python-client